refactor: move all marketing components to folder
