+++ src/app/(marketing)/knowledge-hub/components/blog.tsx
@@ -2,7 +2,7 @@
 import { BlogsConfig } from '@/types';
 import Image from 'next/image';
 import { faCalendarDays, faClock, faUser } from '@fortawesome/free-solid-svg-icons';
-import TextWithIcon from '@/components/text-with-icon';
+import TextWithIcon from '@/components/marketing/text-with-icon';
 
 export default function BlogPage({ blog }: { blog: BlogsConfig }) {
   const dateString = blog.date.toLocaleDateString('en-AU');
