+++ src/components/marketing/main-section-image.tsx
