+++ src/components/marketing/rotating-word.tsx
