+++ src/config/nav.ts
@@ -1,11 +1,21 @@
 import { NavConfig } from '@/types';
 
-export const navConfig: NavConfig = {
+export const marketingNavConfig: NavConfig = {
   navItems: [
     { title: 'About', href: '/about' },
     { title: 'Knowledge Hub', href: '/knowledge-hub' },
     { title: 'Enterprise Solutions', href: '/enterprise-solutions' },
     { title: 'Legal', href: '/legal' },
     { title: 'Get in Touch', href: '/get-in-touch' },
+  ],
+};
+
+export const applicatonNavConfig: NavConfig = {
+  navItems: [
+    { title: 'Home', href: '/home' },
+    { title: 'Development Hub', href: '/development-hub' },
+    { title: 'My Skills', href: '/my-skills' },
+    { title: 'Messages', href: '/messages' },
+    { title: 'Notifications', href: '/notifications' },
   ],
 };
